+++ packages/slate-plugins/src/common/__tests__/transforms/mergeDeepToNodes/descendant-editor-elements-nested.spec.tsx
@@ -4,7 +4,7 @@ import { jsx } from '../../../../__test-utils__/jsx';
 import { ELEMENT_LI } from '../../../../elements/list/defaults';
 import { ELEMENT_PARAGRAPH } from '../../../../elements/paragraph/index';
 import { isDescendant } from '../../../queries/index';
-import { setPropsToNodes } from '../../../transforms/index';
+import { mergeDeepToNodes } from '../../../transforms/index';
 
 const node = (
   <editor>
@@ -30,8 +30,12 @@ const output = (
 ) as any;
 
 it('should set props to all descendants', () => {
-  setPropsToNodes(node, props, {
-    filter: ([n]) => isDescendant(n),
+  mergeDeepToNodes({
+    node,
+    source: props,
+    query: {
+      filter: ([n]) => isDescendant(n),
+    },
   });
   expect(node.children).toEqual(output.children);
 });