+++ packages/slate-plugins/src/elements/blockquote/index.ts
@@ -1,3 +1,4 @@
+export * from './components';
 export * from './BlockquotePlugin';
 export * from './defaults';
 export * from './deserializeBlockquote';