+++ packages/slate-plugins/src/elements/media-embed/components/index.ts
@@ -1,1 +1,2 @@
+export * from './MediaEmbedElement.styles';
 export * from './MediaEmbedElement';