+++ packages/slate-plugins/src/elements/todo-list/components/index.ts
@@ -1,1 +1,2 @@
+export * from './TodoListElement.styles';
 export * from './TodoListElement';